backend/requirements.txt
flask
flask-cors
yt-dlp
openai
openai-whisper
faster-whisper
mlx-whisper
pyannote.audio
requests
gunicorn
gevent
python-dotenv

pytest
pytest-mock
pytest-cov
